// erxes-40ba19a7b455493283ac0585c7068f630adee6df/ui/src/modules/settings/integrations/types.ts
export interface IMessengerData {
  botEndpointUrl?: string;
  botShowInitialMessage?: boolean;
  messages?: IMessages;
  notifyCustomer?: boolean;
  supporterIds?: string[];
  availabilityMethod?: string;
  isOnline?: boolean;
  timezone?: string;
  requireAuth?: boolean;
  showChat?: boolean;
  showLauncher?: boolean;
  forceLogoutWhenResolve?: boolean;
  showVideoCallRequest?: boolean;
  onlineHours?: IOnlineHour[];
  links?: ILink;
}

export interface IUiOptions {
  color?: string;
  textColor?: string;
  wallpaper?: string;
  logo?: string;
  logoPreviewUrl?: string;
}

export interface ITopic {
  topicId: string;
}

export interface IWebsite {
  url: string;
  buttonText: string;
  description: string;
}

export interface ILead {
  formCode: string;
}

export interface ITopicMessengerApp {
  credentials: ITopic;
}

export interface IWebsiteMessengerApp {
  credentials: IWebsite;
}

export interface ILeadMessengerApp {
  credentials: ILead;
}
export interface IMessengerApps {
  knowledgebases?: ITopic[];
  websites?: IWebsite[];
  leads?: ILead[];
}

export interface IIntegration {
  _id: string;
  kind: string;
  name: string;
  brandId?: string;
  code: string;
  formId: string;
  languageCode?: string;
  createUrl: string;
  createModal: string;
  messengerData?: IMessengerData;
  form: IForm;
  uiOptions?: IUiOptions;
  leadData: ILeadData;
  brand: IBrand;
  channels: IChannel[];
  isActive?: boolean;
  webhookData?: IWebhookData;
  leadMessengerApps?: ILeadMessengerApp[];
  websiteMessengerApps?: IWebsiteMessengerApp[];
  knowledgeBaseMessengerApps?: ITopicMessengerApp[];
}

export interface IAccount {
  _id: string;
  name: string;
  kind: string;
  id: string;
}

// query types
export type QueryVariables = {
  page?: number;
  perPage?: number;
  searchValue?: string;
};

export type IntegrationTypes =
  | 'facebook'
  | 'gmail'
  | 'nylas-gmail'
  | 'nylas-exchange'
  | 'nylas-imap'
  | 'nylas-office365'
  | 'nylas-outlook'
  | 'nylas-yahoo'
  | 'twitter'
  | 'smooch-telegram'
  | 'smooch-viber'
  | 'smooch-line'
  | 'smooch-twilio';

export type IntegrationsQueryResponse = {
  integrations: IIntegration[];
  loading: boolean;
  refetch: (variables?: QueryVariables) => Promise<any>;
};